requirements.txt
fastapi
uvicorn
openai
requests
huggingface-hub
python-dotenv